image refactored

+++ controllers/imageController.js
@@ -4,13 +4,14 @@ const sharp = require('sharp');
 const Image = require('../modals/imageModal');
 const removeFile = require('../utils/removeFile');
 const catchAsync = require('../utils/catchAsync');
+const getFileUrl = require('../utils/getFileUrl');
+const ApiFeatures = require('../utils/apiFeatures');
+const AppError = require('../utils/CustomError');
 
 const filePath = path.join(`${__dirname}/../uploads`);
 
 exports.createImage = catchAsync(async (req, res, next) => {
-  const originalName = req.file.originalname.split('.')[0];
-  const filename = `${Date.now()}-${originalName}.webp`;
-  const imgUrl = `http://${req.get('host')}/static/images/${filename}`;
+  const { filename, imgUrl } = getFileUrl(req);
 
   // upload and save file to webp format
   await sharp(req.file.buffer)
@@ -33,11 +34,26 @@ exports.createImage = catchAsync(async (req, res, next) => {
   });
 });
 
-exports.getImages = catchAsync(async (req, res, next) => {
-  const images = await Image.find();
+exports.getAllImages = catchAsync(async (req, res, next) => {
+  // const images = await Image.find();
+  let images = new ApiFeatures(Image.find(), req.query)
+    .filter()
+    .sort()
+    .fields()
+    .pagination();
+  images = await images.query;
   return res.status(200).json({
     status: 'success',
+    results: images.length,
     images,
+  });
+});
+
+exports.getImage = catchAsync(async (req, res, next) => {
+  const image = await Image.findById(req.params.id);
+  res.status(200).json({
+    status: 'success',
+    image,
   });
 });
 
@@ -54,13 +70,14 @@ exports.searchImage = catchAsync(async (req, res, next) => {
 
 exports.deleteImage = catchAsync(async (req, res, next) => {
   const image = await Image.findByIdAndDelete(req.params.id);
+  if (!image) return next(new AppError('Image not found', 404));
   const name = image.imageUrl.split('/');
 
   // remove image from the directory
   await removeFile(path.join(`${__dirname}/../uploads`, name[name.length - 1]));
 
-  res.status(200).json({
+  res.status(204).json({
     status: 'success',
-    image,
+    data: null,
   });
 });
